Refactor - remove CellFace class from media.h/cpp to new media_cellface.h/cpp.
